Move PythonInterpreterGuard
